+++ database/migrations/2019_10_22_190000_add_is_pinned_to_posts_table.php
@@ -1,0 +1,22 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddIsPinnedToPostsTable extends Migration
+{
+    public function up()
+    {
+        Schema::table('posts', function (Blueprint $table) {
+            $table->boolean('is_pinned')->index()->default(false)->after('is_private');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('posts', function (Blueprint $table) {
+            $table->dropColumn('is_pinned');
+        });
+    }
+}